refactor player feature UI to player_uimodule ref:#49
