+++ .idea/misc.xml
@@ -66,6 +66,8 @@
         <entry key="../../../../layout/compose-model-1650037421730.xml" value="0.55" />
         <entry key="../../../../layout/compose-model-1650112608895.xml" value="0.1" />
         <entry key="../../../../layout/compose-model-1650113204507.xml" value="0.22677364864864866" />
+        <entry key="../../../../layout/compose-model-1650256720869.xml" value="0.335" />
+        <entry key="../../../../layout/compose-model-1650257451404.xml" value="0.375" />
         <entry key="app/src/main/res/layout/layout_main.xml" value="0.11820652173913043" />
       </map>
     </option>
