+++ app/build.gradle
@@ -65,6 +65,7 @@ dependencies {
     implementation 'androidx.activity:activity-compose:1.3.1'
     implementation project(path: ':ui-core')
     implementation project(path: ':ui-home')
+    implementation project(path: ':ui-player')
     testImplementation 'junit:junit:4.+'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
