+++ app/src/main/java/com/imdmp/youtubecompose/MainAppScreen.kt
@@ -17,15 +17,15 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
+import com.imdmp.ui_player.VideoPlayerScreen
+import com.imdmp.ui_player.controls.ControlsCallback
 import com.imdmp.youtubecompose.base.ui.navigation.model.Destination
 import com.imdmp.youtubecompose.features.fullscreenview.FullScreenView
 import com.imdmp.youtubecompose.features.profile.ProfileScreen
 import com.imdmp.youtubecompose.features.splash.SplashScreen
 import com.imdmp.youtubecompose.features.videolist.VideoListScreen
 import com.imdmp.youtubecompose.features.videolist.model.VideoListViewModel
-import com.imdmp.youtubecompose.features.videoplayer.VideoPlayerScreen
-import com.imdmp.youtubecompose.features.videoplayer.controls.ControlsCallback
-import com.imdmp.youtubecompose.features.videoplayer.model.VideoPlayerViewModel
+import com.imdmp.youtubecompose.features.videoplayer.VideoPlayerViewModel
 import timber.log.Timber
 
 @Composable
