--- app/src/main/java/com/imdmp/youtubecompose/features/videoplayer/VideoEvent.kt
@@ -1,11 +1,0 @@
-package com.imdmp.youtubecompose.features.videoplayer
-
-sealed class VideoEvent {
-    object ToggleStatus : VideoEvent()
-
-    object VideoLoaded : VideoEvent()
-
-    object VideoError: VideoEvent()
-
-    object VideoLoading:VideoEvent()
-}
