--- app/src/main/java/com/imdmp/youtubecompose/features/videoplayer/comments/CommentState.kt
@@ -1,17 +1,0 @@
-package com.imdmp.youtubecompose.features.videoplayer.comments
-
-data class CommentState(
-    val commentModelList: List<CommentModel>,
-    val isLoading: Boolean,
-    val error: Error? = null
-) {
-
-}
-
-data class CommentModel(
-    val authorName: String,
-    val commentText: String,
-    val profilePic: String,
-    val likeCount: Int,
-    val timeCommented: String,
-)
