--- app/src/main/java/com/imdmp/youtubecompose/features/videoplayer/controls/ControlsCallback.kt
@@ -1,19 +1,0 @@
-package com.imdmp.youtubecompose.features.videoplayer.controls
-
-interface ControlsCallback {
-    fun fullScreenClicked()
-    fun pauseOrPlayClicked()
-
-    companion object{
-        fun default(): ControlsCallback = object : ControlsCallback {
-            override fun fullScreenClicked() {
-                TODO("Not yet implemented")
-            }
-
-            override fun pauseOrPlayClicked() {
-                TODO("Not yet implemented")
-            }
-
-        }
-    }
-}
