--- app/src/main/java/com/imdmp/youtubecompose/features/videoplayer/fullscreenmode/VideoPlayerFullScreen.kt
@@ -1,22 +1,0 @@
-package com.imdmp.youtubecompose.features.videoplayer.fullscreenmode
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import com.imdmp.youtubecompose.BaseActivityCallbacks
-
-@Composable
-fun VideoPlayerFullScreen(baseActivityCallbacks: BaseActivityCallbacks) {
-    Box(
-        Modifier
-            .fillMaxSize()
-            .background(Color.Black)
-    ) {
-
-    }
-
-
-}
