+++ ui-core/src/main/res/drawable/food13.webp
