+++ ui-player/src/androidTest/java/com/imdmp/ui_player/player/PlaybackTest.kt
@@ -1,0 +1,67 @@
+package com.imdmp.ui_player.player
+
+import dagger.hilt.android.testing.HiltAndroidTest
+import org.junit.Test
+
+@HiltAndroidTest
+class PlaybackTest {
+
+//    @get:Rule(order = 0)
+//    var hiltAndroidRule = HiltAndroidRule(this)
+//
+//    @get:Rule(order = 1)
+//    val composeTestRule = createAndroidComposeRule<MainActivity>()
+//
+//    @BindValue
+//    var splashViewModel: SplashViewModel = mock()
+//
+//    @Mock
+//    lateinit var videoPlayerScreenCallbacks: VideoPlayerScreenCallbacks
+//
+//    @Mock
+//    lateinit var mockedNavController: NavController
+//
+//    private lateinit var exoPlayer: ExoPlayer
+//
+//    private lateinit var context: Context
+//
+//    @UiThreadTest
+//    @Before
+//    fun setUp() {
+//        MockitoAnnotations.openMocks(this)
+//        context = ApplicationProvider.getApplicationContext()
+//        exoPlayer = TestExoPlayerBuilder(context).build()
+//
+//        hiltAndroidRule.inject()
+//    }
+//
+//    @Test
+//    fun Init_And_PlayVideo_When_Playback_Init() {
+//        val dummyUrlString = "www.sampleurl"
+//        composeTestRule.setContent {
+//            Playback(
+//                player = exoPlayer,
+//                playerScreenCallbacks = videoPlayerScreenCallbacks
+//            )
+//        }
+//
+//        verify(videoPlayerScreenCallbacks).prepareAndPlayVideoPlayer(dummyUrlString)
+//    }
+
+    @Test
+    fun Player_Control_SelectFullScreen_FullScreenDisplay() {
+//        composeTestRule.setContent {
+//            val navController = rememberNavController()
+//            MainAppScreen(navController = navController)
+//            Playback(
+//                playerScreenCallbacks = videoPlayerScreenCallbacks,
+//                player = exoPlayer
+//            )
+//        }
+//
+//
+//        composeTestRule.onNodeWithTag(Tags.TAG_BUTTON_SET_FULLSCREENVIEW).performClick()
+//        composeTestRule.onNodeWithTag(Destination.FullScreenView.path).assertExists()
+
+    }
+}
