+++ ui-player/src/main/java/com/imdmp/ui_player/Tags.kt
@@ -1,0 +1,10 @@
+package com.imdmp.ui_player
+
+object Tags {
+    const val TEST = "test"
+    const val TAG_FULLSCREENVIEW = "TAG_FULLSCREEN_VIEW"
+    const val TAG_BUTTON_SET_FULLSCREENVIEW ="TAG_BUTTON_SET_FULLSCREENVIEW"
+    const val TAG_COMMENTS_LIST = "TAG_COMMENTS_LIST"
+    const val TAG_PLAYER_LIFECYCLER_HANDLER ="TAG_PLAYER_LIFECYCLER_HANDLER"
+    const val TAG_PAUSE_PLAY_BUTTON ="TAG_PAUSE_PLAY_BUTTON"
+}
