+++ ui-player/src/main/java/com/imdmp/ui_player/VideoEvent.kt
@@ -1,0 +1,11 @@
+package com.imdmp.ui_player
+
+sealed class VideoEvent {
+    object ToggleStatus : VideoEvent()
+
+    object VideoLoaded : VideoEvent()
+
+    object VideoError: VideoEvent()
+
+    object VideoLoading: VideoEvent()
+}
