+++ ui-player/src/main/java/com/imdmp/ui_player/fullscreenmode/VideoPlayerFullScreen.kt
@@ -1,0 +1,22 @@
+package com.imdmp.ui_player.fullscreenmode
+
+//import androidx.compose.foundation.background
+//import androidx.compose.foundation.layout.Box
+//import androidx.compose.foundation.layout.fillMaxSize
+//import androidx.compose.runtime.Composable
+//import androidx.compose.ui.Modifier
+//import androidx.compose.ui.graphics.Color
+//import com.imdmp.youtubecompose.BaseActivityCallbacks
+
+//@Composable
+//fun VideoPlayerFullScreen(baseActivityCallbacks: BaseActivityCallbacks) {
+//    Box(
+//        Modifier
+//            .fillMaxSize()
+//            .background(Color.Black)
+//    ) {
+//
+//    }
+//
+//
+//}
