+++ ui-player/src/main/java/com/imdmp/ui_player/playback/PlaybackScreenCallbacks.kt
@@ -1,0 +1,20 @@
+package com.imdmp.ui_player.playback
+
+interface PlaybackScreenCallbacks {
+    fun prepareAndPlayVideoPlayer(url: String = "")
+
+    fun disposeVideoPlayer()
+
+    companion object {
+        fun default(): PlaybackScreenCallbacks = object : PlaybackScreenCallbacks {
+            override fun prepareAndPlayVideoPlayer(url: String) {
+                TODO("Not yet implemented")
+            }
+
+            override fun disposeVideoPlayer() {
+                TODO("Not yet implemented")
+            }
+
+        }
+    }
+}
